+++ src/components/IndexPage/Greeting/Greeting.module.css
@@ -1,0 +1,55 @@
+.Greeting {
+  min-height: calc(100vh - var(--ifm-navbar-height));
+  overflow: hidden;
+  position: relative;
+  padding: 0 2rem;
+}
+
+.Content {
+  max-width: var(--max-content-width);
+  margin: auto;
+  padding-top: 6rem;
+  position: relative;
+}
+
+.Text {
+  width: 66.666666666%;
+  position: relative;
+}
+
+.Title {
+  font-size: 4rem;
+  margin-bottom: 2rem;
+  hyphens: none;
+  word-wrap: normal;
+  max-width: 48rem;
+}
+
+.Description {
+  font-size: 1.5rem;
+  margin-bottom: 2rem;
+}
+
+.Animation {
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  left: 0;
+  width: 100vw;
+  transform: translate(0, calc(100% - 5rem));
+}
+
+@media (max-width: 1000px) {
+  .Title {
+    font-size: 2.8rem;
+  }
+  .Text {
+    width: unset;
+  }
+}
+
+@media (max-width: 800px) {
+  .Animation {
+    bottom: 0;
+  }
+}
