+++ src/components/IndexPage/Greeting/SineWaves/SineWavesAnimation.module.css
@@ -1,0 +1,15 @@
+.SineWavesAnimation {
+  position: relative;
+  transform: translate(20%, 0);
+}
+
+.SineWavesAnimation::before {
+  content: '';
+  display: block;
+  position: absolute;
+  height: 4px;
+  right: 100%;
+  bottom: calc(50% + 2px);
+  width: 100%;
+  background-color: #198fff;
+}
