+++ src/css/custom.css
@@ -1,13 +1,3 @@
-/* stylelint-disable docusaurus/copyright-header */
-/**
- * Any CSS included here will be global. The classic template
- * bundles Infima by default. Infima is a CSS framework designed to
- * work well for content-centric websites.
- */
-
-/* WAVES WRAPPER */
-
-
 /* PILL BUTTON */
 .pill-btn,
 article
