--- src/css/navbar.css
@@ -1,7 +1,0 @@
-.dropdown__link.navbar__link--active {
-  color: #1c1e21;
-}
-
-.navbar {
-  box-shadow: 0 0 2px rgba(0, 0, 0, 0.2);
-}
