+++ src/css/variables.css
@@ -1,6 +1,12 @@
-:root {
+:root,
+:host {
+  font-size: 16px;
+  --max-content-width: 1200px;
+
+  /* Infima is a CSS framework that Docusaurus uses by default.
+     https://github.com/facebook/docusaurus/issues/3955
+  */
   --ifm-font-family-base: "Red Hat Display", Arial, San-serif;
-  --ifm-font-size-base: 16px;
   --ifm-color-primary: #198fff;
   --ifm-color-primary-dark: #1379da;
   --ifm-color-primary-darker: #0e65b6;
@@ -13,7 +19,7 @@
   --ifm-container-width-xl: 1536px;
   --ifm-navbar-background-color: #fff;
   --ifm-navbar-link-color: #198fff;
-  --ifm-navbar-shadow: none;
+  --ifm-navbar-shadow: 0 0 2px rgba(0, 0, 0, 0.2);
   --ifm-footer-background-color: #464e56;
   --ifm-hero-background-color: #fff;
   --bg-dark-grey: #464e56;
@@ -22,4 +28,3 @@
   --ifm-navbar-height: 62px;
   --ifm-navbar-item-padding-vertical: 13px;
 }
-
