+++ src/pages/community.tsx
@@ -1,3 +1,3 @@
-import CommunityPage from '@site/src/components/CommunityPage';
+import CommunityPage from '@site/src/components/CommunityPage/CommunityPage';
 
 export default CommunityPage;
