+++ src/pages/index.tsx
@@ -1,0 +1,3 @@
+import IndexPage from "../components/IndexPage/IndexPage";
+
+export default IndexPage;
